requirements.txt
python-dotenv==1.0.1
gradio==4.19.0
langchain==0.1.11
openai==1.13.3
faiss-cpu==1.7.4
langchain-community==0.0.27
=========

aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
appdirs==1.4.4
argcomplete==3.1.1
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
azure-ai-formrecognizer==3.3.2
azure-common==1.1.28
azure-core==1.30.1
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
bs4==0.0.2
build==1.0.3
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.1
contourpy==1.2.0
crewai==0.30.11
cryptography==42.0.2
cycler==0.12.1
dataclasses-json==0.6.4
dataclasses-json-speakeasy==0.5.11
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
distro==1.9.0
dnspython==2.5.0
docstring-parser==0.15
effdet==0.4.1
embedchain==0.1.103
emoji==2.10.1
en-core-web-lg @ file:///C:/Users/319533/OneDrive%20-%20Cognizant/Desktop/MouseWithoutBorders/en_core_web_lg-3.6.0-py3-none-any.whl#sha256=9012584a8fcf5bc61bedb13a4b511fb81ffff9f770958ee06366671be10632e9
et-xmlfile==1.1.0
executing==2.0.1
faiss-cpu==1.7.4
fastapi==0.109.2
ffmpy==0.3.2
filelock==3.13.1
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
google-api-core==2.19.0
google-auth==2.27.0
google-cloud-aiplatform==1.51.0
google-cloud-bigquery==3.23.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
gptcache==0.1.43
gradio==4.19.0
gradio_client==0.10.0
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.63.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.6
importlib-metadata==7.0.0
importlib-resources==6.1.1
instructor==0.5.2
iopath==0.1.10
ipykernel==6.29.2
ipython==8.21.0
isodate==0.6.1
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.1.11
langchain-community==0.0.27
langchain-core==0.1.52
langchain-experimental==0.0.51
langchain-openai==0.0.5
langchain-text-splitters==0.0.1
langchainhub==0.1.14
langdetect==1.0.9
langgraph==0.0.48
langsmith==0.1.23
layoutparser==0.3.4
linkedin_api==2.1.1
lxml==5.1.0
Mako==1.3.5
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
msrest==0.7.1
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.1
openai==1.13.3
opencv-python==4.8.0.76
openpyxl==3.1.2
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.9.14
overrides==7.7.0
packaging==23.2
pandas==2.2.0
parso==0.8.3
pathy==0.10.3
pdf2image==1.17.0
pdfminer.six==20221105
pdfplumber==0.10.4
pikepdf==8.11.0
pillow==10.2.0
pillow_heif==0.15.0
pipx==1.2.0
platformdirs==4.2.0
plum-dispatch==1.7.4
portalocker==2.8.2
posthog==3.4.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.23.4
psutil==5.9.8
pulsar-client==3.4.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pydub==0.25.1
Pygments==2.17.2
pymongo==4.6.1
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
pynvml==11.5.0
pypandoc==1.12
pyparsing==3.0.9
pypdf==4.0.1
pypdfium2==4.27.0
PyPika==0.48.9
pyproject_hooks==1.0.0
pyreadline3==3.4.1
pysbd==0.3.4
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.1
python-iso639==2024.2.7
python-magic==0.4.27
python-multipart==0.0.9
python-pptx==0.6.23
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
rapidfuzz==3.6.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.18.0
rsa==4.9
ruff==0.2.1
safetensors==0.3.2
schema==0.7.7
scipy==1.10.1
semantic-version==2.10.0
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
spacy==3.6.1
Spire.Xls==14.2.2
SQLAlchemy==2.0.30
stack-data==0.6.3
starlette==0.36.3
sympy==1.12
tabulate==0.9.0
tavily-python==0.3.1
tenacity==8.2.3
tiktoken==0.5.2
timm==0.9.12
tokenizers==0.15.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.0
torchvision==0.17.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.37.1
typer==0.9.0
types-requests==2.31.0.20240125
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
unstructured==0.12.6
unstructured-client==0.18.0
unstructured-inference==0.7.23
unstructured.pytesseract==0.3.12
urllib3==1.26.18
userpath==1.8.0
uvicorn==0.27.0.post1
watchfiles==0.21.0
wcwidth==0.2.13
websocket-client==1.7.0
websockets==11.0.3
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.1.9
yarl==1.9.4
zipp==3.17.0